libusbgx: add USBG_VERSION preprocessor symbol
The libusbgx API has changed over time (for example struct
usbg_gadget_strs as changed from fixed-size fields to pointers to char)
which requires that client programs also change.

Sometimes it is not possible to keep applications and libusbgx versions
in lock step and it is desirable to support multiple libusbgx API
versions from a single client code base.

Add a version number which can be used for conditional code in client
applications by doing:

	#if USBG_VERSION < USBG_MAKE_VERSION(0, 2, 0)

Signed-off-by: John Keeping <john@metanate.com>
[Add macro for generating API version]
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>