libusbgx: Introduce helper macro for casting to union
Casting some variable to union is a gcc extension that is not
available in g++. As we are using this feature in our headers
this causes compilation issues in C++ code.

First attempt to fix this took place when we were merging UAC2
support. But used solution is far away from being perfect as
it does not provide any type safety so is basically equivalent
of casting to void *.

To fix this let's introduce a macro that will cast any variable
to union using {} initialization which is supported by both gcc
and g++ and additionally provides us type safety.

Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>