libusbgx: uac2: Use dedicated macro for casting to union
Casting to union is gcc extension that is not supported
by g++. As we are using it in header in causes compilation
issues. Let's fix this by using our own macro for type-safe
casting to union.

Initially we used *(union a)&var construction but it does
not provide any type safety so let's replace it with our
macro.

Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>