libusbgx: tests: Add a C++ compilation checker
Recently we faced some C++ compilation issues because
we used in our header a gcc extension that is not supported
by g++. Even through this problem has been fix let's try to
avoid such problems in a future. To do so, let's add a C++
compilation check for all our public headers as part of our
test suit.

Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>